ruff as formatter & linter, fixed et json result

+++ src/nt25/lib/et.py
@@ -1,4 +1,6 @@
+import copy
 import os
+from sys import intern
 import time
 import json
 import struct
@@ -9,16 +11,22 @@ from datetime import UTC, datetime, timedelta, timezone
 from PIL import Image as pi
 from exif import Image, DATETIME_STR_FORMAT
 
-VERSION = '0.2.2'
+VERSION = "0.2.3"
 MAX_WIDTH = 1000
+THUMBNAIL_WIDTH = 352
+
 COMMENT_SEGMENT = b"\xff\xfe"
 EPOCH = datetime.fromtimestamp(0, UTC)
-IMAGE_EXT = ('.jpg', '.jpeg', '.png', '.bmp')
+IMAGE_EXT = (".jpg", ".jpeg", ".png", ".bmp")
+
+
+NOT_FOUND = "not found"
+INTERNAL_FAILED = "internal failed"
 
 
 def dms2dec(dms: tuple):
   d, m, s = dms
-  return d + m/60 + s/3600
+  return d + m / 60 + s / 3600
 
 
 def dtFormatter(str):
@@ -36,26 +44,33 @@ def gpsDt2Dt(date, time, offset=8):
 
 
 def optimizeFolder(dir, q=80, mw=MAX_WIDTH):
-  list = []
+  total = 0
+  shrink = 0
 
   for d, _, f in os.walk(dir):
     for file in f:
       _, e = os.path.splitext(file)
       if e.lower() in IMAGE_EXT:
         path = os.path.join(d, file)
+        total += 1
 
         if optimizeFile(path, q=q, mw=mw):
-          list.append(path)
+          shrink += 1
 
-  return len(list)
+  return {"total": total, "shrink": shrink}
 
 
-def optimizeFile(file, q=80, mw=MAX_WIDTH):
-  less = False
+def genThumbnail(file):
+  _, ext = os.path.splitext(file)
+  name = file[: -len(ext)] + ".thumbnail" + ext
+  return optimizeFile(file, mw=THUMBNAIL_WIDTH, copyExif=False, name=name)
+
+
+def optimizeFile(file, q=80, mw=MAX_WIDTH, copyExif=True, name=None):
+  optimized = False
 
   if not os.path.exists(file):
-    print("not found:", file)
-    return less
+    return optimized
 
   img = pi.open(file)
   w, h = img.size
@@ -67,25 +82,39 @@ def optimizeFile(file, q=80, mw=MAX_WIDTH):
     h = int(scale * h)
     img = img.resize((w, h))
 
-  _, e = os.path.splitext(file)
-  if img.mode == 'RGB':
-    e = '.jpg'
+  _, ext = os.path.splitext(file)
 
-  ofile = file + e.lower()
+  end = ext.lower()
+  if img.mode == "RGB":
+    end = ".jpg"
 
-  if e == '.png':
-    img.save(ofile, optimize=True, compress_level=9)
+  ofile = file + end
+
+  if end == ".png":
+    img.save(ofile, optimize=True)
   else:
-    img.save(ofile, quality=q, optimize=True, progression=True)
+    img.save(ofile, quality=80, optimize=True)
+
+  img.close()
 
   if os.path.getsize(ofile) < os.path.getsize(file) * 0.8:
-    less = True
-    transplant(file, ofile)
-    os.replace(ofile, file)
-  else:
-    os.remove(ofile)
+    optimized = True
 
-  return less
+    if copyExif:
+      transplant(file, ofile, optimize=False)
+
+    if name is not None:
+      os.replace(ofile, name)
+    else:
+      os.replace(ofile, file)
+
+  else:
+    if name is not None:
+      os.replace(ofile, name)
+    else:
+      os.remove(ofile)
+
+  return optimized
 
 
 def tryGet(img, key, default):
@@ -100,68 +129,70 @@ def tryGet(img, key, default):
 
 
 def dumpExif(file, optimize=False):
-  result = {}
+  result = {"version": VERSION}
 
   if not os.path.exists(file):
+    result["file"] = NOT_FOUND
     return result
 
-  if optimize:
-    optimizeFile(file)
-
-  with open(file, 'rb') as f:
+  with open(file, "rb") as f:
     img = Image(f)
     for key in img.get_all():
       try:
         result[key] = str(img[key])
       except Exception:
+        # result["file"] = INTERNAL_FAILED
         pass
+
+  if optimize:
+    result["optimized"] = optimizeFile(file)  # type: ignore
 
   return result
 
 
 def parseExif(file, optimize=False):
+  result = {}
+
   if not os.path.exists(file):
-    return {}
+    result["file"] = NOT_FOUND
+    return result
+
+  with open(file, "rb") as f:
+    img = Image(f)
 
   if optimize:
-    optimizeFile(file)
+    result["optimized"] = optimizeFile(file)
 
-  with open(file, 'rb') as f:
-    try:
-      img = Image(f)
-    except Exception:
-      return {}
-
-  width = tryGet(img, 'pixel_x_dimension', -1)
-  height = tryGet(img, 'pixel_y_dimension', -1)
+  width = tryGet(img, "pixel_x_dimension", -1)
+  height = tryGet(img, "pixel_y_dimension", -1)
 
   if width < 0:
-    width = tryGet(img, 'image_width', -1)
-    height = tryGet(img, 'image_height', -1)
+    width = tryGet(img, "image_width", -1)
+    height = tryGet(img, "image_height", -1)
 
-  create = tryGet(img, 'datetime_original', None)
-  modify = tryGet(img, 'datetime', None)
+  create = tryGet(img, "datetime_original", None)
+  modify = tryGet(img, "datetime", None)
 
   createDt = None if create is None else dtFormatter(create)
   modifyDt = None if modify is None else dtFormatter(modify)
 
-  latitude = tryGet(img, 'gps_latitude', None)
+  latitude = tryGet(img, "gps_latitude", None)
   latitude = None if latitude is None else dms2dec(latitude)
 
-  latRef = tryGet(img, 'gps_latitude_ref', default='N')
-  if latRef != 'N' and latitude:
+  latRef = tryGet(img, "gps_latitude_ref", default="N")
+  if latRef != "N" and latitude:
     latitude = -latitude
 
-  longitude = tryGet(img, 'gps_longitude', None)
+  longitude = tryGet(img, "gps_longitude", None)
   longitude = None if longitude is None else dms2dec(longitude)
 
-  longRef = tryGet(img, 'gps_longitude_ref', default='E')
-  if longRef != 'E' and longitude:
+  longRef = tryGet(img, "gps_longitude_ref", default="E")
+  if longRef != "E" and longitude:
     longitude = -longitude
 
   gpsDatetime = None
-  gd = tryGet(img, 'gps_datestamp', None)
-  gt = tryGet(img, 'gps_timestamp', None)
+  gd = tryGet(img, "gps_datestamp", None)
+  gt = tryGet(img, "gps_timestamp", None)
 
   if gd and gt:
     offset = int(time.localtime().tm_gmtoff / 3600)
@@ -178,25 +209,29 @@ def parseExif(file, optimize=False):
     offset = None
     offsetDelta = None
 
-  return {
-      'width': width,
-      'height': height,
-      'latitude': latitude,
-      'longitude': longitude,
-      'datetime.create': dt2str(createDt),
-      'datetime.modify': dt2str(modifyDt),
-      'datetime.gps': dt2str(gpsDatetime),
-      'ts': ts,
-      'offset': offset,
-      'offset.delta': offsetDelta,
-  }
+  result.update(
+    {
+      "width": width,
+      "height": height,
+      "latitude": latitude,
+      "longitude": longitude,
+      "datetime.create": dt2str(createDt),
+      "datetime.modify": dt2str(modifyDt),
+      "datetime.gps": dt2str(gpsDatetime),
+      "ts": ts,
+      "offset": offset,
+      "offset.delta": offsetDelta,
+    }
+  )
+
+  return result
 
 
 class InvalidImageDataError(ValueError):
   pass
 
 
-def genSegments(data):
+def _segments(data):
   if data[0:2] != b"\xff\xd8":
     return []
 
@@ -204,31 +239,31 @@ def genSegments(data):
   segments = [b"\xff\xd8"]
 
   while 1:
-    if data[head: head + 2] == b"\xff\xda":
+    if data[head : head + 2] == b"\xff\xda":
       segments.append(data[head:])
       break
 
     else:
-      length = struct.unpack(">H", data[head + 2: head + 4])[0]
+      length = struct.unpack(">H", data[head + 2 : head + 4])[0]
       endPoint = head + length + 2
-      seg = data[head: endPoint]
+      seg = data[head:endPoint]
       segments.append(seg)
       head = endPoint
 
-    if (head >= len(data)):
+    if head >= len(data):
       raise InvalidImageDataError("Wrong JPEG data.")
 
   return segments
 
 
-def setComment(segments, comment: str, enc='utf-8'):
+def _comment(segments, comment: str, enc="utf-8"):
   contains = False
 
   if len(segments) > 1:
     cb = comment.encode(enc)
     length = len(cb) + 2
 
-    cbSeg = COMMENT_SEGMENT + length.to_bytes(2, byteorder='big') + cb
+    cbSeg = COMMENT_SEGMENT + length.to_bytes(2, byteorder="big") + cb
 
     for i in range(len(segments)):
       if segments[i][0:2] == COMMENT_SEGMENT:
@@ -243,15 +278,7 @@ def setComment(segments, comment: str, enc='utf-8'):
   return segments
 
 
-def getComment(segments, enc='utf-8'):
-  for seg in segments:
-    if seg[0:2] == COMMENT_SEGMENT:
-      return seg[4:].decode(encoding=enc, errors='replace')
-
-  return None
-
-
-def getExif(segments):
+def _exif(segments):
   for seg in segments:
     if seg[0:2] == b"\xff\xe1" and seg[4:10] == b"Exif\x00\x00":
       return seg
@@ -259,11 +286,13 @@ def getExif(segments):
   return b""
 
 
-def mergeSegments(segments, exif=b""):
+def _merge(segments, exif=b""):
   if len(segments) > 1:
-    if segments[1][0:2] == b"\xff\xe0" and \
-            segments[2][0:2] == b"\xff\xe1" and \
-            segments[2][4:10] == b"Exif\x00\x00":
+    if (
+      segments[1][0:2] == b"\xff\xe0"
+      and segments[2][0:2] == b"\xff\xe1"
+      and segments[2][4:10] == b"Exif\x00\x00"
+    ):
       if exif:
         segments[2] = exif
         segments.pop(1)
@@ -276,9 +305,7 @@ def mergeSegments(segments, exif=b""):
       if exif:
         segments[1] = exif
 
-    elif (segments[1][0:2] == b"\xff\xe1" and
-          segments[1][4:10] == b"Exif\x00\x00"):
-
+    elif segments[1][0:2] == b"\xff\xe1" and segments[1][4:10] == b"Exif\x00\x00":
       if exif:
         segments[1] = exif
       elif exif is None:
@@ -292,109 +319,153 @@ def mergeSegments(segments, exif=b""):
 
 
 def removeExif(file, optimize=False):
+  result = {}
+
   if not os.path.exists(file):
-    print('bad file:', file)
-    return
+    result["file"] = NOT_FOUND
+    return result
 
-  if optimize:
-    optimizeFile(file)
-
-  with open(file, 'rb') as f:
+  with open(file, "rb") as f:
     data = f.read()
 
-  segments = genSegments(data)
-  segments = list(filter(lambda seg: not (seg[0:2] == b"\xff\xe1"
-                                          and seg[4:10] == b"Exif\x00\x00"),
-                         segments))
+  segments = _segments(data)
 
-  segments = setComment(segments, "nt25.et")
-  data = b"".join(segments)
+  if len(segments) > 1:
+    segments = list(
+      filter(
+        lambda seg: not (seg[0:2] == b"\xff\xe1" and seg[4:10] == b"Exif\x00\x00"),
+        segments,
+      )
+    )
 
-  with open(file, "wb+") as f:
-    f.write(data)
+    segments = _comment(segments, "nt25.et")
+    data = b"".join(segments)
+
+    with open(file, "wb+") as f:
+      f.write(data)
+      result["result"] = True
+
+      if optimize:
+        result["optimized"] = optimizeFile(file)
+
+  return result
 
 
 def transplant(src, dst, optimize=False):
+  result = {}
+
   if not os.path.exists(src) or not os.path.exists(dst):
-    print(f'not found: {src}, {dst}')
-    return
+    result["file"] = NOT_FOUND
+    return result
 
-  if optimize:
-    optimizeFile(dst)
-
-  with open(src, 'rb') as f:
+  with open(src, "rb") as f:
     s = f.read()
 
-  segments = genSegments(s)
-  exif = getExif(segments)
+  segments = _segments(s)
+  exif = _exif(segments)
 
-  with open(dst, 'rb') as f:
-    d = f.read()
+  if optimize:
+    result["optimized"] = optimizeFile(dst, copyExif=False)
 
-  segments = genSegments(d)
+  with open(dst, "rb") as f:
+    data = f.read()
+
+  segments = _segments(data)
 
   if len(segments) > 1:
-    segments = setComment(segments, "nt25.et")
-    d = mergeSegments(segments, exif)
+    segments = _comment(segments, "nt25.et")
+    data = _merge(segments, exif)
 
     with open(dst, "wb+") as f:
-      f.write(d)
+      f.write(data)
+      result["result"] = True
+
+  return result
 
 
 def main():
   parser = argparse.ArgumentParser(description="EXIF tool")
-  parser.add_argument('-v', '--version', action='store_true',
-                      help='echo version')
-  parser.add_argument('-o', '--optimize', action='store_true',
-                      help='optimize jpg file, work with -r, -d, -c, -f')
-  parser.add_argument('-r', '--rm', action='store_true',
-                      help='remove meta, use: -r -f FILE')
-  parser.add_argument('-d', '--dump', action='store_true',
-                      help='dump meta, use: -d -f FILE')
-  parser.add_argument('-c', '--copy', type=str,
-                      help='copy meta, use: -c SRC -f DST')
-  parser.add_argument('-f', '--file', type=str, help='image file')
-  parser.add_argument('-m', '--max', type=int, default=MAX_WIDTH,
-                      help='set max width when shrink file')
-  parser.add_argument('-s', '--shrink', type=str,
-                      help=f'shrink file/folder with max width {MAX_WIDTH}')
+  parser.add_argument("-v", "--version", action="store_true", help="echo version")
+  parser.add_argument("-f", "--file", type=str, help="image file")
+  parser.add_argument(
+    "-d", "--dump", action="store_true", help="dump meta, use: -d -f FILE"
+  )
+  parser.add_argument("-c", "--copy", type=str, help="copy meta, use: -c SRC -f DST")
+  parser.add_argument(
+    "-r", "--rm", action="store_true", help="remove meta, use: -r -f FILE"
+  )
+  parser.add_argument(
+    "-o",
+    "--optimize",
+    action="store_true",
+    help="optimize jpg file, work with -r, -d, -c, -f",
+  )
+
+  parser.add_argument(
+    "-s",
+    "--shrink",
+    type=str,
+    help=f"shrink file/folder with max width {MAX_WIDTH}",
+  )
+  parser.add_argument(
+    "-m",
+    "--max",
+    type=int,
+    default=MAX_WIDTH,
+    help="set max width when shrinking",
+  )
+  parser.add_argument(
+    "-t",
+    "--thumbnail",
+    action="store_true",
+    help="generate thumbnail when shrinking",
+  )
 
   args = parser.parse_args()
 
+  result = {}
+
   if args.version:
-    print(f'et: {VERSION}')
-    return
+    result = {"version": VERSION}
+
+  if args.file is not None:
+    opt = True if args.optimize else False
+
+    if args.dump:
+      r = dumpExif(args.file, optimize=opt)
+    elif args.copy:
+      r = transplant(args.copy, args.file, optimize=opt)
+    elif args.rm:
+      r = removeExif(args.file, optimize=opt)
+    else:
+      r = parseExif(args.file, optimize=opt)
+
+    result.update(r)  # type: ignore
 
   if args.shrink is not None:
-    if os.path.isfile(args.shrink):
-      r = optimizeFile(args.shrink, mw=args.max)
-    elif os.path.isdir(args.shrink):
-      r = optimizeFolder(args.shrink, mw=args.max)
+    max = args.max if args.max is not None else MAX_WIDTH
+    if os.path.isdir(args.shrink):
+      r = optimizeFolder(args.shrink, mw=max)
     else:
-      r = 404
+      r = {"result": optimizeFile(args.shrink, mw=max)}
 
-    print(f'{{"less": {r} }}')
-    return
+    if args.thumbnail:
+      r.update({"thumbnail": genThumbnail(args.shrink)})
 
-  if args.file is None:
-    print('usage: et [-h] [-v] [-o] [-f FILE] [-d -f FILE]\n'
-          '\t\t[-r -f FILE] [-c SRC -f DST]\n'
-          '\t\t[[-m MAX] -s FILE]')
-    return
+    result.update(r)  # type: ignore
 
-  opt = True if args.optimize else False
-
-  if args.dump:
-    r = dumpExif(args.file, optimize=opt)
-  elif args.rm:
-    r = removeExif(args.file, optimize=opt)
-  elif args.copy:
-    r = transplant(args.copy, args.file, optimize=opt)
+  if len(result) > 0:
+    print(
+      json.dumps(
+        result,
+        indent=2,
+      )
+    )
   else:
-    r = parseExif(args.file, optimize=opt)
-
-  if r is not None:
-    print(json.dumps(r, indent=2, sort_keys=False))
+    print(
+      "usage: et [-h] [-v] [-f FILE [-d] [-o] [-r]]\n"
+      "\t\t[-c SRC -f DST] [-s FILE [-t] [-m MAX]]"
+    )
 
 
 if __name__ == "__main__":
